refactor(models): use dynamic import for subcommand
